Update README.md
Credit Card Fraud can be defined as a case where a person uses someone else’s credit card for personal reasons while the owner and the card-issuing authorities are unaware of the fact that the card is being used. Due to the rise and acceleration of E-Commerce, there has been a tremendous use of credit cards for online shopping which led to a high amount of fraud related to credit cards. In the era of digitalization, the need to identify credit card frauds is necessary. Fraud detection involves monitoring and analyzing the behavior of various users in order to estimate detect or avoid undesirable behavior. In order to identify credit card fraud detection effectively, we need to understand the various technologies, algorithms, and types involved in detecting credit card frauds. The algorithm can differentiate transactions that are fraudulent or not. Find fraud, they need to pass the dataset and knowledge of the fraudulent transactions. They analyze the dataset and classify all transactions. Fraud detection involves monitoring the activities of populations of users in order to estimate, perceive or avoid objectionable behavior, which consists of fraud, intrusion, and defaulting. Machine learning algorithms are employed to analyze all the authorized transactions and report the suspicious ones. These reports are investigated by professionals who contact the cardholders to confirm if the transaction was genuine or fraudulent. So, here I have an Anonymized credit card transactions raw dataset. First I imported all the important libraries and then uploaded the data set. at first, I analyzed the data nature and meaning of every variable. here time of transaction is in sec format so I converted it into the date-time format, In the data set Class represents the type of transaction, it is a genuine transaction or fraud transaction. here class--> 0 represent Genuine transaction and Class--> 1 represent Fraud transaction. At first Exploratory Data Analysis is done and through analysis, I analyzed that the data set is highly Unbalanced. Hence model will give the wrong prediction if the data will be unbalanced so first I balanced the data and then did the Model Preparation. I have analyzed this Data by Random Forest Model and then Gradient boosting Model to detect fraud transaction.